api/requirements.txt
Flask==3.1.2
gunicorn==23.0.0
Werkzeug==3.1.3
requests==2.32.5
flask-cors==6.0.1
astropy==7.1.0
google-cloud-firestore==2.21.0